+++ src/file.c
@@ -3,20 +3,20 @@
 #include "general.h"
 #include "file.h"
 
-size_t file_len(const char name[static 1]) {
+long file_len(const char name[static 1]) {
     struct stat res;
     stat(name, &res);
     return res.st_size;
 }
 
-int readFile(void *dst, const char name[static 1], unsigned max_len) {
-    size_t len = file_len(name);
+int readFile(void *dst, const char name[static 1], long max_len) {
+    long len = file_len(name);
     if (len > max_len) {
-        fprintf(stderr, "File too long: %lu > %u\n", len, max_len);
+        fprintf(stderr, "File too long: %ld > %ld\n", len, max_len);
         return -1;
     }
 
-    DBG_PRINT("File size = %lu\n"
+    DBG_PRINT("File size = %ld\n"
               "Opening...\n", len);
 
     FILE *file = fopen(name, "rb");
@@ -27,7 +27,7 @@ int readFile(void *dst, const char name[static 1], unsigned max_len) {
 
     DBG_PRINT("Opened. Reading...\n");
 
-    int read_bytes = fread(dst, 1, len, file);
+    int read_bytes = (int)fread(dst, 1, (unsigned long)len, file);
     fclose(file);
 
     DBG_PRINT("Read %d bytes\n", read_bytes);
